refactor some methods into utils.py
